splitted invite tests
